+++ samples/hello/hello_table.rb
@@ -302,15 +302,16 @@ class HelloTable
         }
         
         # Data-bind table items (rows) to a model collection (BaseballGame.schedule),
-        # mapping columns in declaration order to row model properties (attributes)
-        # By convention, every column property can be accompanied by extra properties
-        # with the following suffixes: `_background`, `_foreground`, `_font`, and `_image`
-        # For example, for `game_date`, model could also implement these related properties:
+        # automatically inferring model attributes from column names by convention
+        # (e.g. 'Home Team' column assumes a `home_team` attribute on models)
+        #
+        # By convention, every inferred model attribute can be accompanied by extra
+        # model attributes to set extra table properties with the following suffixes:
+        # `_background`, `_foreground`, `_font`, and `_image`
+        #
+        # For example, for :game_date, model could also implement these related properties:
         # `game_date_background`, `game_date_foreground`, `game_date_font`, `game_date_image`
-        # That is done in order to let the table widget set extra properties if needed.
-        items <=> [BaseballGame, :schedule,
-                    column_properties: [:game_date, :game_time, :ballpark, :home_team, :away_team, :promotion]
-                  ]
+        items <=> [BaseballGame, :schedule]
         
         # Data-bind table selection
         selection <=> [BaseballGame, :selected_game]